+++ src/main/java/tf/tailfriend/pet/message/SuccessMessage.java
@@ -6,7 +6,8 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public enum SuccessMessage {
-    PET_FOUND_SUCCESS("반려동물 정보 조회에 성공하였습니다");
+    PET_FOUND_SUCCESS("반려동물 정보 조회에 성공하였습니다"),
+    GET_FRIENDS_SUCCESS("친구 조회에 성공하였습니다");
 
     private final String message;
 }
