+++ src/main/java/tf/tailfriend/user/exception/UserSaveException.java
@@ -3,9 +3,10 @@ package tf.tailfriend.user.exception;
 import org.springframework.http.HttpStatus;
 import tf.tailfriend.global.exception.CustomException;
 
-import static tf.tailfriend.user.message.error.ErrorMessage.USER_SAVE_FAIL;
+import static tf.tailfriend.user.message.ErrorMessage.USER_SAVE_FAIL;
 
 public class UserSaveException extends CustomException {
+
     @Override
     public HttpStatus getStatus() {
         return HttpStatus.INTERNAL_SERVER_ERROR;
